+++ src/main/java/com/wardk/meeteam_backend/web/notification/dto/NotificationResponse.java
@@ -1,0 +1,33 @@
+package com.wardk.meeteam_backend.web.notification.dto;
+
+
+import com.wardk.meeteam_backend.domain.notification.entity.Notification;
+import com.wardk.meeteam_backend.domain.notification.entity.NotificationType;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDate;
+
+@Data
+public class NotificationResponse {
+
+    private Long id;
+    private NotificationType type;
+    private String message;
+    private boolean isRead;
+    private LocalDate createdAt;
+
+    private NotificationPayLoad payload;
+
+
+    public NotificationResponse(Notification notification) {
+        this.id = notification.getId();
+        this.type = notification.getType();
+        this.message = notification.getMessage();
+        this.isRead = notification.isRead();
+        this.createdAt = LocalDate.from(notification.getCreatedAt());
+        this.payload = new NotificationPayLoad(notification.getProject(), notification.getReceiver());
+    }
+
+}
